packages/core/src/database/user.ts
import { Insertable, Selectable, Updateable } from 'kysely'
import { AuditedEntity } from "./audited_entity";


export interface UserTable extends AuditedEntity {
  first_name: string;
  last_name: string;
  title?: string;
    photo?: string;
    invite_status?: number;
    super_admin?: boolean;
    company_id: number;
}

// You should not use the table schema interfaces directly. Instead, you should
// use the `Selectable`, `Insertable` and `Updateable` wrappers. These wrappers
// make sure that the correct types are used in each operation.
export type User = Selectable<UserTable>
export type NewUser = Insertable<UserTable>
export type UserUpdate = Updateable<UserTable>
